Extracted unit test base TensorFlatSpecBase in the JVM integer algebra unit test suite
