+++ jvm-algebra/test/src/integer/JVMArrayTensorSpec.scala
@@ -6,19 +6,7 @@ import scala.reflect.ClassTag
 import bmaso.tensoralg.abstractions._
 import org.scalatest.{FlatSpec, Matchers}, Matchers._
 
-abstract class JVMArrayTensorSpecBase[T: Numeric: ClassTag: ConvertsIntArray] extends FlatSpec {
-  val numeric: Numeric[T] = implicitly[Numeric[T]]
-
-  implicit def canConvertToInt(t: T) = new {
-    def asI: Int = numeric.toInt(t)
-  }
-
-  implicit def canConvertIntArray(array: Array[Int]) = new {
-    def asTArray: Array[T] = {
-      implicitly[ConvertsIntArray[T]].convertIntArray(array)
-    }
-  }
-
+abstract class JVMArrayTensorSpecBase[T: Numeric: ClassTag: ConvertsIntArray] extends TensorFlatSpecBase[T] {
   "A 5x5 tensor" should "have expected magnitude, order, and elementSize" in {
     val t = ArrayTensor[T](((0 to 24) toArray).asTArray, Array(5, 5), 0)
     t.order should be (2)
